draft

+++ src/main/java/ru/kors/springstudents/repository/StudentRepository.java
@@ -7,9 +7,9 @@ import java.util.List;
 
 
 public interface StudentRepository extends JpaRepository<Student, Long> {
-    void deleteByEmail(int id);
+    void deleteByEmail(long id);
 
-    Student findStudentById(int id);
+    Student findStudentById(long id);
 
     Student save(Student student);
